fix(rtsp): correct the package doc's UDP claim and unbreak 32-bit test builds
doc.go stated "This client speaks only the TCP-interleaved profile; there
is no UDP transport", which has been false since the UDP work landed:
PreferUDP, PreferUDPThenTCP, ErrUDPSetupRejected, setupUDP and six UDP
test files all exist. It is the package overview, so it was the first
paragraph a pkg.go.dev reader saw. Replace it with a description of both
transports, and generalize the routing paragraph that assumed every frame
arrives interleaved on the connection reader goroutine.

transport_test.go's TestParseSession table typed wantTOSec as int while
its largest-representable row holds 9223372036, so the rtsp test binary
did not compile for GOARCH=arm or 386. Non-test code was unaffected, and
CI only runs amd64, but the release workflow ships an armv7 binary, so
the suite could not be run on a platform this module distributes for.
Widen the field to int64; the conversion at the use site already handles
it.

Verified with build, vet (amd64, arm64), test -race, golangci-lint and
gofmt, plus test-binary compilation for arm/v7 and 386.

+++ rtsp/transport_test.go
@@ -180,10 +180,13 @@ func TestBuildTransportRoundTrip(t *testing.T) {
 func TestParseSession(t *testing.T) {
 	t.Parallel()
 	tests := []struct {
-		name      string
-		value     string
-		wantID    string
-		wantTOSec int
+		name   string
+		value  string
+		wantID string
+		// int64, not int: the largest-representable row holds 9223372036,
+		// which overflows a 32-bit int and would not compile on GOARCH=arm
+		// or 386, both of which this module ships binaries for.
+		wantTOSec int64
 	}{
 		{"bare ID default timeout", "12345678", "12345678", 60},
 		{"explicit timeout", "12345678;timeout=60", "12345678", 60},
